pyproject.toml
[project]
name = "multi-agent-coding-system"
version = "0.1.0"
description = "A multi-agent AI coder. Reached #12 on Stanford's TerminalBench"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "terminal-bench",
    "litellm>=1.72.6",
    "pyyaml>=6.0.2",
    "pydantic>=2.11.5",
]

[tool.setuptools]
packages = ["src"]


[tool.uv.sources]
multi-agent-coding-system = { workspace = true }
terminal-bench = { git = "https://github.com/Danau5tin/terminal-bench-py311.git", branch = "main" }
